Create Kid Guess Game App
The idea behind this game to make use of importation of Python Module (Random) and create a guess game for a kid.
Play will enter a number at a prompt to guess the random. The if statement is used to control the output of each of input from the player who has a total 5 lives at each prompt played.